final
So, the game is rubbish =) and you need an xbox controller to play. I can bring one in if you don't have one already.
theres 3 flying modes. horizontal, vertical, and twinstick which changes as you enter new waves. added an enemy. theres a skybox that moves based on the flying mode. I had a plan for this game when I did all this last sunday, but in the past week I had no interest in working on it. Anyway, its better than nothing